Create system to require users for specific views.
This uses ideas from
https://drstearns.github.io/tutorials/gomiddleware/

The high-level idea is to save information about the user to the session
store and pull it out in middleware and make it available to views. The
views that need a user then require it via the function parameters so
it's can't be silently ignored. We then need an adapter between regular
views and our special views that require users.