Switch to an interface-based system for dealing with fs data
I've struggled with the exact right way to represent FieldSeeker-owned
data schema and the layer that we are building above it. Since I'm so
new to Go the obvious answer, interfaces, came slowly. I'm finally
getting it. This doesn't yet transform everything into interfaces as I
eventually will, but it starts the pattern with geometry/latlong data.

Also remove the now-unused notes "query" with mocked data.