Rework types to use implementation-hiding structs
So, the interfaces don't work. That's because in order for the pgx row
scanner to do its job transforming a database row into a struct that
struct needs to have *public* properties that correspond to the column
names. That means I can't hide the data fields in those structs.

Instead, I'm now creating structs that wrap those structs, with methods
that access the underlying data. This makes it so the rest of the code
can interact with nice methods and be unaware of the data underneath.
Much better separation of concerns.

This change is breaking all clients of the API.