Break auth into its own module, rework auth error system
I was getting some odd behavior from the iOS client, and I wanted to
understand better what was going on so I separated out the
authentication logic to its own file and added some custom Error types
so that I could return error codes in WWW-Authenticate-Error. This
allowed me to discover that my session was incorrectly getting the user
ID out of the session, causing it to occasionally claim there was no
session when there actually was.

I'm also responding with a 202 to login when no redirect is provided via
the 'next' queryarg so that the iOS client doesn't screw up when testing
the login.